refactor(brokers): split http_sdk.py and ws_sdk.py into separate class files
- Split HttpService and HttpClient from http_sdk.py into http_service.py and http_client.py
- Split WebSocket and WebSocketClient from ws_sdk.py into ws_service.py and ws_client.py
- Original SDK files now re-export for backward compatibility
- Updated all broker imports to use new split files
- Standardized logger naming to f"{self.__class__.__name__}_{self.name}"
