+++ src/brokers/base/ws_sdk.py
@@ -1,242 +1,6 @@
-# Built-in Library
-from abc import ABC, abstractmethod
-import time
-import threading
-import websocket
+# Backward compatibility re-exports
+# Import from the new split files instead
+from src.brokers.base.ws_service import WebSocket
+from src.brokers.base.ws_client import WebSocketClient
 
-from src.infrastructure.logging.set_logger import get_logger, get_adapter
-from src.core.models.trade import TradePair
-
-logger = get_logger(__name__)
-
-
-class WebSocket(ABC):
-    _instance_counter: int = 1  # shared across all instance
-
-    @classmethod
-    def _generate_id(cls) -> int:
-        cls._instance_counter += 1
-        return (cls._instance_counter - 1)
-
-    def _operation_logging(
-        self,
-        log: str,
-    ) -> None:
-        log = str(log)
-        self.logger.info(log)
-
-    def generate_timestamp(self) -> int:
-        return int(time.time() * 1_000)
-
-    def __init__(
-        self,
-        name: str,
-        api_key: str | None,
-        secret_key: str | None,
-        ping_interval: int,  # in seconds -> every <ping_interval> send the hearbeat
-        url: str | None = None,
-    ) -> None:
-        self.logger = get_adapter(logger, self.__class__.__name__)
-        
-        # wss endpoint
-        self.url: str = url
-
-        # instance specific identifier
-        self._id: int = self._generate_id()
-        self._name: str = name
-
-        # api_key and secret_key
-        self.api_key: str = api_key
-        self.secret_key: str = secret_key
-        self.authenticated: bool | None = None
-
-        # ping interval
-        self.ping_interval: int = ping_interval
-
-        # threads list
-        self.threads: list[threading.Thread] = []
-
-        return
-
-    '''
-    ####################################################################################
-    #                            Instance Specific Field                               #
-    ####################################################################################
-    '''
-    @property
-    def id(self) -> int:
-        return self._id
-
-    @property
-    def name(self) -> str:
-        return self._name
-
-    '''
-    ####################################################################################
-    #                                Abstract Method                                   #
-    ####################################################################################
-    '''
-
-    @abstractmethod
-    def connect(self) -> None:
-        '''
-        ;func connect()
-            - connect to the endpoint
-        '''
-        return
-
-    @abstractmethod
-    def disconnect(self) -> None:
-        '''
-        ;func disconnect()
-            - close WebSocket connection gracefully.
-        '''
-        return
-
-    @abstractmethod
-    def subscribe(self) -> None:
-        '''
-        ;func subscribe()
-            - subscribe to the topic/data stream
-        '''
-        return
-
-    @abstractmethod
-    def unsubscribe(self) -> None:
-        '''
-        ;func unsubscribe()
-            - unsubscribe from the topic/data stream
-        '''
-        return
-
-    @abstractmethod
-    def send(self) -> None:
-        '''
-        ;func send()
-            - send message to the server/endpoint
-        '''
-        return
-
-    @abstractmethod
-    def _reconnect(self) -> None:
-        '''
-        ;func _reconnect()
-            - reconnect to the endpoint once it is closed
-        '''
-        return
-
-    '''
-    ####################################################################################
-    #                       Abstract Method - WebSocketApp                             #
-    ####################################################################################
-    '''
-    @abstractmethod
-    def on_open(
-        self,
-        ws: websocket.WebSocketApp,
-    ) -> None:
-        return
-
-    @abstractmethod
-    def on_message(
-        self,
-        ws: websocket.WebSocketApp,
-        msg: str | bytes,
-    ) -> None:
-        return
-
-    @abstractmethod
-    def on_close(
-        self,
-        ws: websocket.WebSocketApp,
-        status_cod: int,
-        close_msg: int,
-    ) -> None:
-        return
-
-    @abstractmethod
-    def on_error(
-        self,
-        ws: websocket.WebSocketApp,
-        error: Exception,
-    ) -> None:
-        return
-
-
-class WebSocketClient(ABC):
-    '''
-    ;class WebSocketClient
-        - Base Class for WebSocket Client for each broker
-        - It defines the contract "each WebSocket for Crypto Broker should have"
-
-    ;TradePair List to check the availabilty? -> how to keep updating the TradePair?
-    '''
-    def __init__(
-        self,
-        name: str,
-    ) -> None:
-        self.name: str = name
-        return
-
-    @classmethod
-    @abstractmethod
-    def _parse_trade_pair(cls, trade_pair: TradePair) -> str:
-        '''
-        ;func _parse_trade_pair()
-            - parse the Trade Pair and return the str data type appropriate for each broker
-            - e.g., for MEXC: BTCUSDT, for Binance: BTC_USDT
-        '''
-        return
-
-    '''
-    ####################################################################################
-    #                                 Market Stream                                    #
-    ####################################################################################
-
-    - ticker for the given symbol
-    - kline for the given symbol
-    - depth for the given symbol
-    '''
-    @abstractmethod
-    def ticker(self) -> None:
-        '''
-        ;func ticker
-            - getting the latest price, best bids and asks for the given contract
-            - real time baseis
-        '''
-        return
-
-    @abstractmethod
-    def kline(self) -> None:
-        return
-
-    @abstractmethod
-    def order_book(self) -> None:
-        '''
-        ;func depth()
-            - return the depth of the market, i.e., Order Book Depth
-            - The entire collection of Bids and Asks for the given contract, organized by price.
-
-        ;The market's ability to sustain relatively large market orders without impacting the price of the security.
-        ;The list of all pending limit orders waiting to be executed.
-        '''
-        return
-
-    '''
-    ####################################################################################
-    #                                 User Stream                                      #
-    ####################################################################################
-
-    - Asset
-    - Order History
-    '''
-
-    '''
-    ####################################################################################
-    #                                    Trade                                         #
-    ####################################################################################
-
-    - Current order
-    - Make Order
-    - Cancel Order
-    '''
+__all__ = ["WebSocket", "WebSocketClient"]
