+++ src/brokers/mexc/http_gateway.py
@@ -4,7 +4,7 @@ import json
 
 # Custom libraries
 from src.infrastructure.logging.set_logger import get_logger, get_adapter
-from src.brokers.base.http_sdk import HttpService
+from src.brokers.base.http_service import HttpService
 
 logger = get_logger(__name__)
 
@@ -27,7 +27,7 @@ class MexcFutureGateway(HttpService):
             secret_key = secret_key,
             base_url = base_url,
         )
-        self.logger = get_adapter(logger, self.name)
+        self.logger = get_adapter(logger, f"{self.__class__.__name__}_{self.name}")
         # Set the specific content type for MEXC
         self.set_content_type("application/json")
